Extract string_concatenation_builtin_function in own file
string_builtin_function is too large.

Concatenation functions can live in the
string_concatenation_builtin_function file.

Doing this for all string functions.
