+++ src/solvers/strings/string_builtin_function.cpp
@@ -37,21 +37,6 @@ string_insertion_builtin_functiont::string_insertion_builtin_functiont(
   args.insert(args.end(), fun_args.begin() + 5, fun_args.end());
 }
 
-string_concatenation_builtin_functiont::string_concatenation_builtin_functiont(
-  const exprt &return_code,
-  const std::vector<exprt> &fun_args,
-  array_poolt &array_pool)
-  : string_insertion_builtin_functiont(return_code, array_pool)
-{
-  PRECONDITION(fun_args.size() >= 4 && fun_args.size() <= 6);
-  const auto arg1 = expr_checked_cast<struct_exprt>(fun_args[2]);
-  input1 = array_pool.find(arg1.op1(), arg1.op0());
-  const auto arg2 = expr_checked_cast<struct_exprt>(fun_args[3]);
-  input2 = array_pool.find(arg2.op1(), arg2.op0());
-  result = array_pool.find(fun_args[1], fun_args[0]);
-  args.insert(args.end(), fun_args.begin() + 4, fun_args.end());
-}
-
 optionalt<std::vector<mp_integer>> eval_string(
   const array_string_exprt &a,
   const std::function<exprt(const exprt &)> &get_value)
@@ -102,62 +87,6 @@ array_string_exprt
 make_string(const std::vector<mp_integer> &array, const array_typet &array_type)
 {
   return make_string(array.begin(), array.end(), array_type);
-}
-
-std::vector<mp_integer> string_concatenation_builtin_functiont::eval(
-  const std::vector<mp_integer> &input1_value,
-  const std::vector<mp_integer> &input2_value,
-  const std::vector<mp_integer> &args_value) const
-{
-  const auto start_index =
-    args_value.size() > 0 && args_value[0] > 0 ? args_value[0] : mp_integer(0);
-  const mp_integer input2_size(input2_value.size());
-  const auto end_index =
-    args_value.size() > 1
-      ? std::max(std::min(args_value[1], input2_size), start_index)
-      : input2_size;
-
-  std::vector<mp_integer> eval_result(input1_value);
-  eval_result.insert(
-    eval_result.end(),
-    input2_value.begin() + numeric_cast_v<std::size_t>(start_index),
-    input2_value.begin() + numeric_cast_v<std::size_t>(end_index));
-  return eval_result;
-}
-
-string_constraintst string_concatenation_builtin_functiont::constraints(
-  string_constraint_generatort &generator) const
-
-{
-  auto pair = [&]() -> std::pair<exprt, string_constraintst> {
-    if(args.size() == 0)
-      return add_axioms_for_concat(
-        generator.fresh_symbol, result, input1, input2, array_pool);
-    if(args.size() == 2)
-    {
-      return add_axioms_for_concat_substr(
-        generator.fresh_symbol,
-        result,
-        input1,
-        input2,
-        args[0],
-        args[1],
-        array_pool);
-    }
-    UNREACHABLE;
-  }();
-  pair.second.existential.push_back(equal_exprt(pair.first, return_code));
-  return pair.second;
-}
-
-exprt string_concatenation_builtin_functiont::length_constraint() const
-{
-  if(args.size() == 0)
-    return length_constraint_for_concat(result, input1, input2, array_pool);
-  if(args.size() == 2)
-    return length_constraint_for_concat_substr(
-      result, input1, input2, args[0], args[1], array_pool);
-  UNREACHABLE;
 }
 
 optionalt<exprt> string_concat_char_builtin_functiont::eval(
