+++ src/solvers/strings/string_builtin_function.h
@@ -340,36 +340,6 @@ protected:
   }
 };
 
-class string_concatenation_builtin_functiont final
-  : public string_insertion_builtin_functiont
-{
-public:
-  /// Constructor from arguments of a function application.
-  /// The arguments in `fun_args` should be in order:
-  /// an integer `result.length`, a character pointer `&result[0]`,
-  /// a string `arg1` of type refined_string_typet,
-  /// a string `arg2` of type refined_string_typet,
-  /// optionally followed by an integer `start` and an integer `end`.
-  string_concatenation_builtin_functiont(
-    const exprt &return_code,
-    const std::vector<exprt> &fun_args,
-    array_poolt &array_pool);
-
-  std::vector<mp_integer> eval(
-    const std::vector<mp_integer> &input1_value,
-    const std::vector<mp_integer> &input2_value,
-    const std::vector<mp_integer> &args_value) const override;
-
-  std::string name() const override
-  {
-    return "concat";
-  }
-
-  string_constraintst
-  constraints(string_constraint_generatort &generator) const override;
-
-  exprt length_constraint() const override;
-};
 
 /// String creation from other types
 class string_creation_builtin_functiont : public string_builtin_functiont
